remove useless tests and add useful ones
